feat(HACBS-1604): add InternalServicesConfig API
Through this new API it's possible to define an allow list which will
reject requests not coming from an allowed namespace. In addition,
this resource allows to set information about the Tekton catalog to
use. It's now possible to define if the Pipelines are meant to be
found on a namespace or on a Bundle. When not specified, the
namespace where the configuration is in will be used. This should
be the exact namespace where the manager is deployed.

Signed-off-by: David Moreno García <damoreno@redhat.com>