chore(RELEASE-1347): remove manager image update from github workflow
This commit removes the step from the app_interface workflow that
updates the manager image reference in app-interface-deployments. It was
updating it to a redhat-appstudio quay image, which is not valid
anymore, it was updating a file that is not used, and we are going to
handle this with a final pipeline from now on.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>