refactor(RELEASE-1378): remove request from InternalRequest.Spec
We used the InternalRequest.Spec.Request field to support cluster
references for a pipeline to execute via the InternalRequest. This was
not very flexible, and we have since moved to git resolvers. All
resolvers are now supported and we no longer use the Spec.Request field,
so this commit removes the support for the Request field from the
codebase.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>